+++ core/consensus/pow/src/verifier.rs
@@ -50,7 +50,7 @@ pub struct PowVerifier<C, AuthorityId> {
 #[forbid(deprecated)]
 impl<B, C, AuthorityId> Verifier<B> for PowVerifier<C, AuthorityId> where
     B: Block,
-    DigestItemFor<B>: CompatibleDigestItem<B, AuthorityId> + ShardingDigestItem<u32>,
+    DigestItemFor<B>: CompatibleDigestItem<B, AuthorityId> + ShardingDigestItem<u16>,
     C: Send + Sync,
     AuthorityId: Decode + Encode + Clone + Send + Sync,
 {
@@ -92,7 +92,7 @@ fn check_header<B, AccountId>(
     hash: B::Hash,
 ) -> Result<(B::Header, DigestItemFor<B>), String> where
     B: Block,
-    DigestItemFor<B>: CompatibleDigestItem<B, AccountId> + ShardingDigestItem<u32>,
+    DigestItemFor<B>: CompatibleDigestItem<B, AccountId> + ShardingDigestItem<u16>,
     AccountId: Decode + Encode + Clone,
 {
     // pow work proof MUST be last digest item
